+++ don't let go but not needed.py
@@ -1,5 +1,3 @@
-# With a correction already implemented: dont forget to initialize an instance of Class "War"
-
 
 from vikingsClasses import Soldier, Viking, Saxon, War
 import random
@@ -28,18 +26,6 @@ while len(great_war.saxonArmy) > 0 and len(great_war.vikingArmy) > 0:
     print(f"round: {round} // Viking army: {len(great_war.vikingArmy)} warriors",f"and Saxon army: {len(great_war.saxonArmy)} warriors")
     print(great_war.showStatus())
     round += 1
-
-
-
-
-
-
-
-
-    
-
-
-
 
 
 # casino game
